requirements/main.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements/main.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
    #   twisted
automat==24.8.1
    # via twisted
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
constantly==23.10.4
    # via twisted
diskcache==5.6.3
    # via pyautogen
distro==1.9.0
    # via openai
docker==7.1.0
    # via pyautogen
exceptiongroup==1.2.2
    # via anyio
fastjsonschema==2.20.0
    # via nbformat
flaml==2.2.0
    # via pyautogen
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via openai
hyperlink==21.0.0
    # via twisted
idna==3.8
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
incremental==24.7.2
    # via twisted
jiter==0.5.0
    # via openai
jsonschema==4.23.0
    # via nbformat
jsonschema-specifications==2023.12.1
    # via jsonschema
jupyter-core==5.7.2
    # via nbformat
jupytext==1.16.4
    # via waldiez (pyproject.toml)
markdown-it-py==3.0.0
    # via
    #   jupytext
    #   mdit-py-plugins
mdit-py-plugins==0.4.1
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
nbformat==5.10.4
    # via jupytext
numpy==1.26.4
    # via
    #   flaml
    #   pyautogen
openai==1.43.0
    # via pyautogen
packaging==24.1
    # via
    #   jupytext
    #   pyautogen
platformdirs==4.2.2
    # via jupyter-core
pyautogen==0.3.0
    # via waldiez (pyproject.toml)
pydantic==2.8.2
    # via
    #   openai
    #   pyautogen
pydantic-core==2.20.1
    # via pydantic
python-dotenv==1.0.1
    # via pyautogen
pyyaml==6.0.2
    # via jupytext
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   docker
    #   tiktoken
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
setuptools==74.0.0
    # via
    #   incremental
    #   zope-interface
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
termcolor==2.4.0
    # via pyautogen
tiktoken==0.7.0
    # via pyautogen
tomli==2.0.1
    # via
    #   incremental
    #   jupytext
tqdm==4.66.5
    # via openai
traitlets==5.14.3
    # via
    #   jupyter-core
    #   nbformat
twisted==24.7.0
    # via waldiez (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   twisted
urllib3==2.2.2
    # via
    #   docker
    #   requests
zope-interface==7.0.3
    # via twisted
